+++ Exploration/ExplorerStackBuilderFactory.php
@@ -8,7 +8,6 @@ use Heptacom\HeptaConnect\Core\Exploration\Contract\ExplorerStackBuilderFactoryI
 use Heptacom\HeptaConnect\Core\Exploration\Contract\ExplorerStackBuilderInterface;
 use Heptacom\HeptaConnect\Core\Portal\PortalStackServiceContainerFactory;
 use Heptacom\HeptaConnect\Dataset\Base\EntityType;
-use Heptacom\HeptaConnect\Portal\Base\Exploration\ExplorerCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Psr\Log\LoggerInterface;
 
@@ -27,12 +26,11 @@ final class ExplorerStackBuilderFactory implements ExplorerStackBuilderFactoryIn
         $flowComponentRegistry = $this->portalContainerFactory
             ->create($portalNodeKey)
             ->getFlowComponentRegistry();
-        $components = new ExplorerCollection();
 
-        foreach ($flowComponentRegistry->getOrderedSources() as $source) {
-            $components->push($flowComponentRegistry->getExplorers($source));
-        }
-
-        return new ExplorerStackBuilder($components, $entityType, $this->logger);
+        return new ExplorerStackBuilder(
+            $flowComponentRegistry->getExplorers(),
+            $entityType,
+            $this->logger
+        );
     }
 }