+++ Portal/FlowComponentRegistry.php
@@ -7,27 +7,27 @@ namespace Heptacom\HeptaConnect\Core\Portal;
 use Heptacom\HeptaConnect\Portal\Base\Builder\FlowComponent;
 use Heptacom\HeptaConnect\Portal\Base\Emission\EmitterCollection;
 use Heptacom\HeptaConnect\Portal\Base\Exploration\ExplorerCollection;
+use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PackageContract;
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
+use Heptacom\HeptaConnect\Portal\Base\Portal\PackageCollection;
 use Heptacom\HeptaConnect\Portal\Base\Reception\ReceiverCollection;
 use Heptacom\HeptaConnect\Portal\Base\StatusReporting\StatusReporterCollection;
 use Heptacom\HeptaConnect\Portal\Base\Web\Http\HttpHandlerCollection;
 
 class FlowComponentRegistry
 {
-    /**
-     * @var class-string[]|null
-     */
-    private ?array $orderedSources = null;
+    private bool $isLoaded = false;
 
     /**
-     * @param array<class-string, ExplorerCollection>       $sourcedExplorers
-     * @param array<class-string, EmitterCollection>        $sourcedEmitters
-     * @param array<class-string, ReceiverCollection>       $sourcedReceivers
-     * @param array<class-string, StatusReporterCollection> $sourcedStatusReporters
-     * @param array<class-string, HttpHandlerCollection>    $sourcedWebHttpHandlers
-     * @param array<class-string, string[]>                 $flowBuilderFiles
+     * @param array<int, ExplorerCollection>       $sourcedExplorers
+     * @param array<int, EmitterCollection>        $sourcedEmitters
+     * @param array<int, ReceiverCollection>       $sourcedReceivers
+     * @param array<int, StatusReporterCollection> $sourcedStatusReporters
+     * @param array<int, HttpHandlerCollection>    $sourcedWebHttpHandlers
+     * @param array<class-string, string[]>        $flowBuilderFiles
      */
     public function __construct(
+        private PackageCollection $packages,
         private array $sourcedExplorers,
         private array $sourcedEmitters,
         private array $sourcedReceivers,
@@ -38,53 +38,58 @@ class FlowComponentRegistry
     }
 
     /**
+     * @deprecated Parameter $source will be removed
      * @param class-string $source
      */
-    public function getExplorers(string $source): ExplorerCollection
+    public function getExplorers(?string $source = null): ExplorerCollection
     {
-        $this->loadSource($source);
+        $this->loadSource();
 
-        return new ExplorerCollection($this->sourcedExplorers[$source] ?? []);
+        return new ExplorerCollection($this->sourcedExplorers);
     }
 
     /**
+     * @deprecated Parameter $source will be removed
      * @param class-string $source
      */
-    public function getEmitters(string $source): EmitterCollection
+    public function getEmitters(?string $source = null): EmitterCollection
     {
-        $this->loadSource($source);
+        $this->loadSource();
 
-        return new EmitterCollection($this->sourcedEmitters[$source] ?? []);
+        return new EmitterCollection($this->sourcedEmitters);
     }
 
     /**
+     * @deprecated Parameter $source will be removed
      * @param class-string $source
      */
-    public function getReceivers(string $source): ReceiverCollection
+    public function getReceivers(?string $source = null): ReceiverCollection
     {
-        $this->loadSource($source);
+        $this->loadSource();
 
-        return new ReceiverCollection($this->sourcedReceivers[$source] ?? []);
+        return new ReceiverCollection($this->sourcedReceivers);
     }
 
     /**
+     * @deprecated Parameter $source will be removed
      * @param class-string $source
      */
-    public function getStatusReporters(string $source): StatusReporterCollection
+    public function getStatusReporters(?string $source = null): StatusReporterCollection
     {
-        $this->loadSource($source);
+        $this->loadSource();
 
-        return new StatusReporterCollection($this->sourcedStatusReporters[$source] ?? []);
+        return new StatusReporterCollection($this->sourcedStatusReporters);
     }
 
     /**
+     * @deprecated Parameter $source will be removed
      * @param class-string $source
      */
-    public function getWebHttpHandlers(string $source): HttpHandlerCollection
+    public function getWebHttpHandlers(?string $source = null): HttpHandlerCollection
     {
-        $this->loadSource($source);
+        $this->loadSource();
 
-        return new HttpHandlerCollection($this->sourcedWebHttpHandlers[$source] ?? []);
+        return new HttpHandlerCollection($this->sourcedWebHttpHandlers);
     }
 
     /**
@@ -92,64 +97,115 @@ class FlowComponentRegistry
      * contribute flow components to the current portal container. The first item will be the FQCN of the portal class.
      * The supporting portal extension FQCNs will be ordered lexicographically.
      *
+     * @deprecated Method will be removed. Call the corresponding getter directly without any arguments.
+     *
      * @return class-string[]
      */
     public function getOrderedSources(): array
     {
-        $result = $this->orderedSources;
+        return [PortalContract::class];
+    }
 
-        if ($result === null) {
-            $result = \array_unique([
-                ...\array_keys($this->sourcedExplorers),
-                ...\array_keys($this->sourcedEmitters),
-                ...\array_keys($this->sourcedReceivers),
-                ...\array_keys($this->sourcedStatusReporters),
-                ...\array_keys($this->sourcedWebHttpHandlers),
-                ...\array_keys($this->flowBuilderFiles),
-            ]);
-            \usort($result, static function (string $portalClassA, string $portalClassB): int {
-                $aT = (int) \is_a($portalClassA, PortalContract::class, true);
-                $bT = (int) \is_a($portalClassB, PortalContract::class, true);
-
-                if ($aT === $bT) {
-                    return \strcmp($portalClassA, $portalClassB);
-                }
-
-                return $bT <=> $aT;
-            });
-
-            $this->orderedSources = $result;
+    private function loadSource(): void
+    {
+        if ($this->isLoaded) {
+            return;
         }
 
-        return $result;
+        foreach ($this->flowBuilderFiles as $source => $files) {
+            if ($files !== []) {
+                $flowBuilder = new FlowComponent();
+
+                $flowBuilder->reset();
+                $flowBuilder->setDefaultPriority(
+                    $this->getSourcePackage($source)->getDefaultFlowComponentPriority()
+                );
+
+                foreach ($files as $file) {
+                    // prevent access to object context
+                    (static function (string $file): void {
+                        include $file;
+                    })($file);
+                }
+
+                foreach ($flowBuilder->buildExplorers() as $priority => $explorer) {
+                    ($this->sourcedExplorers[$priority] ??= new ExplorerCollection())->push([$explorer]);
+                }
+
+                foreach ($flowBuilder->buildEmitters() as $priority => $emitter) {
+                    ($this->sourcedEmitters[$priority] ??= new EmitterCollection())->push([$emitter]);
+                }
+
+                foreach ($flowBuilder->buildReceivers() as $priority => $receiver) {
+                    ($this->sourcedReceivers[$priority] ??= new ReceiverCollection())->push([$receiver]);
+                }
+
+                foreach ($flowBuilder->buildStatusReporters() as $priority => $statusReporter) {
+                    ($this->sourcedStatusReporters[$priority] ??= new StatusReporterCollection())->push([$statusReporter]);
+                }
+
+                foreach ($flowBuilder->buildHttpHandlers() as $priority => $httpHandler) {
+                    ($this->sourcedWebHttpHandlers[$priority] ??= new HttpHandlerCollection())->push([$httpHandler]);
+                }
+            }
+        }
+
+        $this->flowBuilderFiles = [];
+
+        \ksort($this->sourcedExplorers);
+        \ksort($this->sourcedEmitters);
+        \ksort($this->sourcedReceivers);
+        \ksort($this->sourcedStatusReporters);
+        \ksort($this->sourcedWebHttpHandlers);
+
+        $this->sourcedExplorers = \array_merge(...\array_map(
+            static fn (ExplorerCollection $explorers) => $explorers->asArray(),
+            $this->sourcedExplorers
+        ));
+
+        $this->sourcedEmitters = \array_merge(...\array_map(
+            static fn (EmitterCollection $emitters) => $emitters->asArray(),
+            $this->sourcedEmitters
+        ));
+
+        $this->sourcedReceivers = \array_merge(...\array_map(
+            static fn (ReceiverCollection $receivers) => $receivers->asArray(),
+            $this->sourcedReceivers
+        ));
+
+        $this->sourcedStatusReporters = \array_merge(...\array_map(
+            static fn (StatusReporterCollection $statusReporters) => $statusReporters->asArray(),
+            $this->sourcedStatusReporters
+        ));
+
+        $this->sourcedWebHttpHandlers = \array_merge(...\array_map(
+            static fn (HttpHandlerCollection $webHttpHandlers) => $webHttpHandlers->asArray(),
+            $this->sourcedWebHttpHandlers
+        ));
+
+        $this->isLoaded = true;
     }
 
     /**
      * @param class-string $source
      */
-    private function loadSource(string $source): void
+    private function getSourcePackage(string $source): PackageContract
     {
-        $files = $this->flowBuilderFiles[$source] ?? [];
+        $packages = $this->packages->withoutItems();
 
-        if ($files !== []) {
-            $flowBuilder = new FlowComponent();
+        $packages->push($this->packages->filter(
+            static fn (PackageContract $package): bool => \is_a($package, $source)
+        ));
 
-            $flowBuilder->reset();
+        $sourcePackage = $packages->first();
 
-            foreach ($files as $file) {
-                // prevent access to object context
-                (static function (string $file): void {
-                    include $file;
-                })($file);
-            }
-
-            ($this->sourcedExplorers[$source] ??= new ExplorerCollection())->push($flowBuilder->buildExplorers());
-            ($this->sourcedEmitters[$source] ??= new EmitterCollection())->push($flowBuilder->buildEmitters());
-            ($this->sourcedReceivers[$source] ??= new ReceiverCollection())->push($flowBuilder->buildReceivers());
-            ($this->sourcedStatusReporters[$source] ??= new StatusReporterCollection())->push($flowBuilder->buildStatusReporters());
-            ($this->sourcedWebHttpHandlers[$source] ??= new HttpHandlerCollection())->push($flowBuilder->buildHttpHandlers());
-
-            unset($this->flowBuilderFiles[$source]);
+        if (!$sourcePackage instanceof PackageContract) {
+            throw new \Exception(
+                'Unable to find source package "' . $source . '" in built packages.',
+                1693695453
+            );
         }
+
+        return $sourcePackage;
     }
 }