+++ Portal/ServiceContainerCompilerPass/BuildDefinitionForFlowComponentRegistryCompilerPass.php
@@ -8,16 +8,21 @@ use Heptacom\HeptaConnect\Core\Portal\FlowComponentRegistry;
 use Heptacom\HeptaConnect\Core\Portal\PortalStackServiceContainerBuilder;
 use Heptacom\HeptaConnect\Portal\Base\Emission\EmitterCollection;
 use Heptacom\HeptaConnect\Portal\Base\Exploration\ExplorerCollection;
+use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PackageContract;
+use Heptacom\HeptaConnect\Portal\Base\Portal\PackageCollection;
 use Heptacom\HeptaConnect\Portal\Base\Reception\ReceiverCollection;
 use Heptacom\HeptaConnect\Portal\Base\StatusReporting\StatusReporterCollection;
 use Heptacom\HeptaConnect\Portal\Base\Web\Http\HttpHandlerCollection;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\Compiler\PriorityTaggedServiceTrait;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 
 final class BuildDefinitionForFlowComponentRegistryCompilerPass implements CompilerPassInterface
 {
+    use PriorityTaggedServiceTrait;
+
     /**
      * @param array<string, string[]> $flowBuilderFiles
      */
@@ -35,6 +40,7 @@ final class BuildDefinitionForFlowComponentRegistryCompilerPass implements Compi
         $groupedWebHttpHandlers = $this->getServiceReferencesGroupedBySource($container, PortalStackServiceContainerBuilder::WEB_HTTP_HANDLER_SOURCE_TAG);
 
         $container->setDefinition(FlowComponentRegistry::class, (new Definition(FlowComponentRegistry::class))->setArguments([
+            new Reference(PackageCollection::class),
             $this->groupServices(ExplorerCollection::class, $groupedExplorers),
             $this->groupServices(EmitterCollection::class, $groupedEmitters),
             $this->groupServices(ReceiverCollection::class, $groupedReceivers),
@@ -56,17 +62,57 @@ final class BuildDefinitionForFlowComponentRegistryCompilerPass implements Compi
 
     private function getServiceReferencesGroupedBySource(ContainerBuilder $container, string $tag): array
     {
+        /** @var PackageCollection $packages */
+        $packages = $container->get(PackageCollection::class);
+
         $grouped = [];
-        $serviceIds = $container->findTaggedServiceIds($tag);
+        $serviceIds = $this->findAndSortTaggedServices($tag, $container);
 
-        foreach ($serviceIds as $serviceId => $tagData) {
-            $groupKey = $tagData[0]['source'] ?? null;
+        foreach ($serviceIds as $reference) {
+            $definition = $container->findDefinition((string) $reference);
+            $tagData = $definition->getTag($tag);
 
-            if (\is_scalar($groupKey)) {
-                $grouped[(string) $groupKey][] = new Reference($serviceId);
+            $priority = $tagData[0]['priority'] ?? null;
+
+            if (!\is_int($priority)) {
+                $source = $tagData[0]['source'] ?? null;
+
+                if (!\is_string($source)) {
+                    throw new \Exception(
+                        'Tag "' . $tag . '" of service "' . $reference . '" is missing "source" attribute.',
+                        1693671570
+                    );
+                }
+
+                $sourcePackage = $this->getSourcePackage($packages, $source);
+                $priority = $sourcePackage->getDefaultFlowComponentPriority();
             }
+
+            $grouped[$priority][] = $reference;
         }
 
+        \ksort($grouped);
+
         return $grouped;
+    }
+
+    private function getSourcePackage(PackageCollection $builtPackages, string $source): PackageContract
+    {
+        $packages = $builtPackages->withoutItems();
+
+        $packages->push($builtPackages->filter(
+            static fn (PackageContract $package): bool => \get_class($package) === $source
+        ));
+
+        $sourcePackage = $packages->first();
+
+        if (!$sourcePackage instanceof PackageContract) {
+            throw new \Exception(
+                'Unable to find source package "' . $source . '" in built packages.',
+                1693698154
+            );
+        }
+
+        return $sourcePackage;
     }
 }