+++ Reception/ReceiverStackBuilderFactory.php
@@ -8,7 +8,6 @@ use Heptacom\HeptaConnect\Core\Portal\PortalStackServiceContainerFactory;
 use Heptacom\HeptaConnect\Core\Reception\Contract\ReceiverStackBuilderFactoryInterface;
 use Heptacom\HeptaConnect\Core\Reception\Contract\ReceiverStackBuilderInterface;
 use Heptacom\HeptaConnect\Dataset\Base\EntityType;
-use Heptacom\HeptaConnect\Portal\Base\Reception\ReceiverCollection;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Psr\Log\LoggerInterface;
 
@@ -27,12 +26,11 @@ final class ReceiverStackBuilderFactory implements ReceiverStackBuilderFactoryIn
         $flowComponentRegistry = $this->portalContainerFactory
             ->create($portalNodeKey)
             ->getFlowComponentRegistry();
-        $components = new ReceiverCollection();
 
-        foreach ($flowComponentRegistry->getOrderedSources() as $source) {
-            $components->push($flowComponentRegistry->getReceivers($source));
-        }
-
-        return new ReceiverStackBuilder($components, $entityType, $this->logger);
+        return new ReceiverStackBuilder(
+            $flowComponentRegistry->getReceivers(),
+            $entityType,
+            $this->logger
+        );
     }
 }